% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{combine_variance_partition_results}
\alias{combine_variance_partition_results}
\title{Combines variance partition results from multiple analyses into a single dataframe for comparison.}
\usage{
combine_variance_partition_results(vp_results_list)
}
\arguments{
\item{vp_results_list}{List of varPartResults objects from multiple variance partition analyses.}
}
\value{
Dataframe combining variance explained for each variable and cell type across all analyses,
with a column for cell type and columns for each variable's variance explained.
}
\description{
Combines variance partition results from multiple analyses into a single dataframe for comparison.
}
