% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ctp.R
\name{compute_ctp}
\alias{compute_ctp}
\title{Computes cell type proportions for a given grouping.}
\usage{
compute_ctp(df, group_cols, cell_type_cols)
}
\arguments{
\item{df}{Dataframe containing cell metadata.}

\item{group_cols}{Character vector of columns to group by (e.g., donor ID)}

\item{cell_type_cols}{Column name representing cell type annotations.
If multiple columns are provided, then values will be concatenated.}
}
\value{
Dataframe with cell type proportions and total nuclei counts (raw & z-scored) per grouping.
}
\description{
Computes cell type proportions for a given grouping.
}
