% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{compute_logit_CTP}
\alias{compute_logit_CTP}
\title{Compute logit transformations of cell type abundances.}
\usage{
compute_logit_CTP(sample_ctp, cell_type_col, pseudocount = 0.5)
}
\arguments{
\item{sample_ctp}{Dataframe containing cell type proportions for each sample,}

\item{cell_type_col}{Column name in `sample_ctp` that contains cell type labels.}

\item{pseudocount}{Value to add to zero counts to avoid issues with logit.}
}
\value{
Dataframe of logit-transformed cell type proportions, with rows for each cell type and columns for each sample.
}
\description{
Compute logit transformations of cell type abundances.
}
