% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ctp.R
\name{compute_mean_cell_type_metrics}
\alias{compute_mean_cell_type_metrics}
\title{Computes mean metrics per cell type for a given grouping.}
\usage{
compute_mean_cell_type_metrics(df, group_cols, cell_type_cols, metric_cols)
}
\arguments{
\item{df}{Dataframe containing cell metadata.}

\item{group_cols}{Character vector of columns to group by (e.g., donor ID)}

\item{cell_type_cols}{Column name representing cell type annotations.
If multiple columns are provided, then values will be concatenated.}

\item{metric_cols}{Character vector of metric columns to compute means for.}
}
\value{
Dataframe with mean metrics per cell type and grouping.
}
\description{
Computes mean metrics per cell type for a given grouping.
}
