% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glmm.R
\name{extract_beta_binomial_fitted_values}
\alias{extract_beta_binomial_fitted_values}
\title{Extract fitted values and residuals from a fitted beta-binomial GLMM.}
\usage{
extract_beta_binomial_fitted_values(glmm_model)
}
\arguments{
\item{glmm_model}{Fitted beta-binomial GLMM object for a specific cell type.}
}
\value{
Dataframe containing the observed fractions, fitted fractions, and
residuals for each sample in the model fit, along with the original model data.
}
\description{
Extract fitted values and residuals from a fitted beta-binomial GLMM.
}
