% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glmm.R
\name{extract_beta_binomial_fixed_effects}
\alias{extract_beta_binomial_fixed_effects}
\title{Extract fixed effects from a fitted beta-binomial GLMM for a specific cell type.}
\usage{
extract_beta_binomial_fixed_effects(glmm, cell_type, cell_type_col)
}
\arguments{
\item{glmm}{Fitted beta-binomial GLMM object for a specific cell type.}

\item{cell_type}{Cell type corresponding to the fitted model.}

\item{cell_type_col}{Column name to use for the cell type in the output dataframe.}
}
\value{
Dataframe containing the fixed effects estimates, confidence intervals, and cell type information for the specified model.
}
\description{
Extract fixed effects from a fitted beta-binomial GLMM for a specific cell type.
}
