% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{extract_cell_type_specific_qc_covariates}
\alias{extract_cell_type_specific_qc_covariates}
\title{Extracts cell type-specific QC covariates for a given cell type, scales them, and formats them for variance partition input.}
\usage{
extract_cell_type_specific_qc_covariates(
  sample_ctp,
  cell_type,
  cell_type_col,
  qc_covariate_cols
)
}
\arguments{
\item{sample_ctp}{Dataframe containing cell type proportions and QC covariates for each sample.}

\item{cell_type}{Cell type for which to extract QC covariates.}

\item{cell_type_col}{Column name in `sample_ctp` that contains cell type labels.}

\item{qc_covariate_cols}{Character vector of column names in `sample_ctp` that contain the QC covariates to extract and scale.}
}
\value{
Dataframe containing sample IDs, cell type labels, and scaled QC covariate values for the specified cell type, formatted for variance partition input.
}
\description{
Extracts cell type-specific QC covariates for a given cell type, scales them, and formats them for variance partition input.
}
