% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ctp.R
\name{filter_df}
\alias{filter_df}
\title{Filters dataframe.}
\usage{
filter_df(df, filters = NULL, group_cols = NULL, group_filters = NULL)
}
\arguments{
\item{df}{Dataframe to filter.}

\item{filters}{Character vector of filtering expressions.}

\item{group_cols}{Optional; Character vector of columns to group by before filtering}

\item{group_filters}{Optional; Character vector of filtering expressions to apply within groups.}
}
\value{
Filtered dataframe.
}
\description{
Filters dataframe.
}
