% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glm.R
\name{fit_beta_binomial_glmm}
\alias{fit_beta_binomial_glmm}
\title{Fit beta-binomial GLMM to cell type counts.}
\usage{
fit_beta_binomial_glmm(
  sample_df,
  cell_type,
  cell_type_col,
  fixed_effects,
  random_effects
)
}
\arguments{
\item{sample_df}{Dataframe containing cell type counts and metadata for each sample.}

\item{cell_type}{Cell type for which to fit the model.}

\item{cell_type_col}{Column name in `sample_df` that contains cell type labels}

\item{fixed_effects}{Character vector of column names in `sample_df` to include as fixed effects in the model.}

\item{random_effects}{Character vector of column names in `sample_df` to include as random effects.}
}
\value{
Fitted beta-binomial GLMM object for the specified cell type.
}
\description{
Fit beta-binomial GLMM to cell type counts.
}
