% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glmm.R
\name{fit_many_beta_binomial_glmms}
\alias{fit_many_beta_binomial_glmms}
\title{Fit beta-binomial GLMMs for multiple cell types.}
\usage{
fit_many_beta_binomial_glmms(
  sample_df,
  cell_types,
  cell_type_col,
  fixed_effects,
  random_effects
)
}
\arguments{
\item{sample_df}{Dataframe containing cell type counts and metadata for each sample.}

\item{cell_types}{Character vector of cell types for which to fit models.}

\item{cell_type_col}{Column name in `sample_df` that contains cell type labels.}

\item{fixed_effects}{Character vector of column names in `sample_df` to include as fixed effects in the models.}

\item{random_effects}{Character vector of column names in `sample_df` to include as random effects in the models.}
}
\value{
Named list of fitted beta-binomial GLMM objects for each specified cell type.
}
\description{
Fit beta-binomial GLMMs for multiple cell types.
}
