% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{fit_vp_one_cell_type}
\alias{fit_vp_one_cell_type}
\title{Fits variance partition model for a single cell type, incorporating cell type-specific QC covariates.}
\usage{
fit_vp_one_cell_type(
  sample_ctp,
  sample_metadata,
  formula,
  cell_type,
  cell_type_col,
  qc_covariate_cols,
  pseudocount = 0.5
)
}
\arguments{
\item{sample_ctp}{Dataframe containing cell type proportions and QC covariates for each sample.}

\item{sample_metadata}{Dataframe containing metadata for each sample.}

\item{formula}{Formula specifying the fixed and random effects for the variance partition model.}

\item{cell_type}{Cell type for which to fit the variance partition model.}

\item{cell_type_col}{Column name in `sample_ctp` that contains cell type labels.}

\item{qc_covariate_cols}{Character vector of column names in `sample_ctp` that contain the QC covariates to extract and scale for the specified cell type.}

\item{pseudocount}{Value to add to zero counts to avoid issues with logit.}
}
\value{
Dataframe containing the variance explained for each variable.
}
\description{
Fits variance partition model for a single cell type, incorporating cell type-specific QC covariates.
}
