% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glm.R
\name{fixed_effect_pvalue_barplot}
\alias{fixed_effect_pvalue_barplot}
\title{Generate a bar plot of -log10(p-values) for specified fixed effects across cell types.}
\usage{
fixed_effect_pvalue_barplot(fixed_effects_df, fixed_effects, cell_type_col)
}
\arguments{
\item{fixed_effects_df}{Dataframe containing the fixed effects estimates, confidence intervals, and cell type information for all specified models.}

\item{fixed_effects}{Character vector of fixed effect names for which to generate the bar}

\item{cell_type_col}{Column name in `fixed_effects_df` that contains cell type labels.}
}
\value{
ggplot object representing the bar plot of -log10(p-values) for the
specified fixed effects across cell types, with bars colored by fixed effect and grouped by cell type.
}
\description{
Generate a bar plot of -log10(p-values) for specified fixed effects across cell types.
}
