% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{format_and_run_variance_partition}
\alias{format_and_run_variance_partition}
\title{Formats input data and runs variance partition analysis.}
\usage{
format_and_run_variance_partition(
  sample_ctp,
  sample_metadata,
  cell_type_col,
  formula,
  cell_types = NULL,
  pseudocount = 0.5
)
}
\arguments{
\item{sample_ctp}{Dataframe containing cell type proportions for each sample.}

\item{sample_metadata}{Dataframe containing metadata for each sample.}

\item{cell_type_col}{Column name in `sample_ctp` that contains cell type labels.}

\item{formula}{Formula specifying the fixed and random effects for the variance partition model.}

\item{cell_types}{Optional character vector of cell types to include in the analysis.}

\item{pseudocount}{Value to add to zero counts to avoid issues with logit}
}
\value{
varPartResults object containing the results of the variance partition analysis,
}
\description{
Formats input data and runs variance partition analysis.
}
