% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{format_metadata_for_vp}
\alias{format_metadata_for_vp}
\title{Formats metadata for variance partition, ensuring sample IDs are rownames and ordered to match CTP matrix columns.}
\usage{
format_metadata_for_vp(sample_metadata_df)
}
\arguments{
\item{sample_metadata_df}{Dataframe containing sample metadata, with a column for sample IDs.}
}
\value{
Dataframe with sample IDs as rownames, ordered to match the columns of the CTP matrix.
}
\description{
Formats metadata for variance partition, ensuring sample IDs are rownames and ordered to match CTP matrix columns.
}
