% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctp_correlations.R
\name{get_ctp_by_region}
\alias{get_ctp_by_region}
\title{Retrieves cell type proportions for a specific cell type across different brain regions.}
\usage{
get_ctp_by_region(ctp_df, cell_type, cell_type_col, region_col, donor_col)
}
\arguments{
\item{ctp_df}{Dataframe containing cell type proportions, with columns for cell type, brain region, donor ID, and fraction of nuclei.}

\item{cell_type}{Cell type of interest.}

\item{cell_type_col}{Column of dataframe containing cell type labels.}

\item{region_col}{Column of dataframe containing brain region labels.}

\item{donor_col}{Column of dataframe containing donor IDs.}
}
\value{
wide dataframe of CTP, with rows for each donor and columns for each brain region.
}
\description{
Retrieves cell type proportions for a specific cell type across different brain regions.
}
