% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_from_beta_binomial_glmm.R
\name{get_mode}
\alias{get_mode}
\title{Calculate the mode of a vector, ignoring NA values.}
\usage{
get_mode(x)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
The mode of the vector.
}
\description{
Calculate the mode of a vector, ignoring NA values.
}
