% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ctp.R
\name{load_and_compute_ctp}
\alias{load_and_compute_ctp}
\title{Main workflow to compute cell type proportions and optional metrics from a metadata file.}
\usage{
load_and_compute_ctp(
  cell_metadata_file,
  group_cols,
  cell_type_col,
  out_file = NULL,
  metric_cols = NULL,
  filters = NULL,
  group_filters = NULL
)
}
\arguments{
\item{cell_metadata_file}{Path to the cell metadata file (one row for each cell barcode),
with columns for grouping (e.g., donor ID, brain region), cell type, and optional metrics.}

\item{group_cols}{Character vector of columns to group by (e.g., donor ID)}

\item{cell_type_col}{Column name representing cell type annotations.}

\item{out_file}{Optional; Output file to save the results}

\item{metric_cols}{Optional; Character vector of metric columns to compute means for.}

\item{filters}{Optional; Character vector of filtering expressions.}

\item{group_filters}{Optional; Character vector of filtering expressions to apply within groups.}
}
\value{
Dataframe with cell type proportions and optional metrics.
}
\description{
Main workflow to compute cell type proportions and optional metrics from a metadata file.
}
