% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ctp.R
\name{load_and_extract_sample_metadata}
\alias{load_and_extract_sample_metadata}
\title{Main workflow to extract sample-level metadata from a cell metadata file.}
\usage{
load_and_extract_sample_metadata(
  cell_metadata_file,
  group_cols,
  donor_metadata_cols,
  metadata_cols_to_group,
  filters = NULL,
  out_file = NULL
)
}
\arguments{
\item{cell_metadata_file}{Path to the cell metadata file (one row for each cell barcode),
with columns for grouping (e.g., donor ID, brain region) and other metadata.}

\item{group_cols}{Character vector of columns to group by (e.g., donor ID, brain region).}

\item{donor_metadata_cols}{Character vector of donor-level metadata columns to include in the output.}

\item{metadata_cols_to_group}{Character vector of metadata columns to aggregate by concatenating unique values within each group.}

\item{filters}{Optional; Character vector of filtering expressions to apply to the cell-level metadata before extracting sample-level metadata.}

\item{out_file}{Optional; Output file}
}
\value{
Dataframe with one row per sample (defined by `group_cols`), including the specified donor-level metadata and aggregated metadata columns.
}
\description{
This function reads cell-level metadata from a file, applies optional filters,
and extracts sample-level metadata by grouping based on specified columns.
 It can also aggregate certain metadata columns by concatenating unique values within each group.
  The resulting sample metadata can be saved to an output file if specified.
}
