% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_from_beta_binomial_glmm.R
\name{make_prediction_grid}
\alias{make_prediction_grid}
\title{Generate a prediction grid for a fitted model, varying specified
variables and holding others constant.}
\usage{
make_prediction_grid(model, vary = NULL, n_points = 100)
}
\arguments{
\item{model}{A fitted model object for COUNTS data (e.g., from glmmTMB).}

\item{vary}{A named list of variables to vary in the prediction grid,
where each name corresponds to a variable in the model and each value is
 a vector of values to use for that variable.}

\item{n_points}{Number of points to use for numeric variables that are not
 specified in `vary`. Default is 100.

 @return A data frame containing the prediction grid, with columns for each variable in the model.}
}
\description{
Generate a prediction grid for a fitted model, varying specified
variables and holding others constant.
}
