% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctp_correlations.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Generate heatmap of Spearman corrleations.}
\usage{
plot_correlation_heatmap(
  ctp_df,
  cell_types,
  regions,
  cell_type_col = "annotation",
  region_col = "brain_region_abbreviation_simple",
  donor_col = "donor_external_id",
  metric_col = "fraction_nuclei",
  drop_outliers = TRUE
)
}
\arguments{
\item{ctp_df}{Dataframe containing cell type proportions, with columns for
cell type, brain region, donor ID, and the CTP metric.}

\item{cell_types}{Cell types to use.}

\item{regions}{Vector of brain regions to include.}

\item{cell_type_col}{Column of dataframe containing cell type labels.}

\item{region_col}{Column of dataframe containing brain region labels.}

\item{donor_col}{Column of dataframe containing donor IDs.}

\item{metric_col}{Column of dataframe containing the CTP metric to be analyzed (e.g. fraction of nuclei).}

\item{drop_outliers}{Whether to filter out rows marked as outliers (if any) before computing correlations. Default is TRUE.}
}
\value{
A ggplot object of the correlation heatmap.
}
\description{
Generate heatmap of Spearman corrleations.
}
