% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctp_correlations.R
\name{plot_ctp_region_pair_by_celltype}
\alias{plot_ctp_region_pair_by_celltype}
\title{Generates scatterplots comparing cell type proportions between two specified brain regions for multiple cell types.}
\usage{
plot_ctp_region_pair_by_celltype(
  ctp_df,
  region1,
  region2,
  cell_types,
  cell_type_col,
  region_col,
  donor_col,
  nrow = 1
)
}
\arguments{
\item{ctp_df}{Dataframe containing cell type proportions, with columns for cell type, brain region, donor ID, and fraction of nuclei.}

\item{region1}{First brain region to compare.}

\item{region2}{Second brain region to compare.}

\item{cell_types}{Character vector of cell types to compare.}

\item{cell_type_col}{Column of dataframe containing cell type labels.}

\item{region_col}{Column of dataframe containing brain region labels.}

\item{donor_col}{Column of dataframe containing donor IDs.}

\item{nrow}{Number of rows to use in combined plot layout (default is 1).}
}
\value{
cowplot object of combined scatterplots comparing cell type proportions between the two regions for multiple cell types.
}
\description{
Generates scatterplots comparing cell type proportions between two specified brain regions for multiple cell types.
}
