% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctp_correlations.R
\name{plot_ctp_region_pairs}
\alias{plot_ctp_region_pairs}
\title{Generates scatterplots comparing cell type proportions between all pairs of specified brain regions for a specific cell type.}
\usage{
plot_ctp_region_pairs(
  ctp_df,
  regions,
  cell_type,
  cell_type_col,
  region_col,
  donor_col,
  nrow = 2
)
}
\arguments{
\item{ctp_df}{Dataframe containing cell type proportions, with columns for cell type, brain region, donor ID, and fraction of nuclei.}

\item{regions}{Character vector of brain regions to compare.}

\item{cell_type}{Cell type of interest.}

\item{cell_type_col}{Column of dataframe containing cell type labels.}

\item{region_col}{Column of dataframe containing brain region labels.}

\item{donor_col}{Column of dataframe containing donor IDs.}

\item{nrow}{Number of rows to use in combined plot layout (default is 2).}
}
\value{
cowplot object of combined scatterplots comparing cell type proportions between all pairs of specified brain regions.
}
\description{
Generates scatterplots comparing cell type proportions between all pairs of specified brain regions for a specific cell type.
}
