% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_ratios.R
\name{plot_donor_ctp_ratio_between_two_regions}
\alias{plot_donor_ctp_ratio_between_two_regions}
\title{Scatterplot of donor CTP ratio between two regions.}
\usage{
plot_donor_ctp_ratio_between_two_regions(
  ratio_df,
  region1,
  region2,
  ratio = NULL
)
}
\arguments{
\item{ratio_df}{Dataframe containing CTP ratios for each donor and brain region,
as generated by `generate_ctp_ratio_table()`.}

\item{region1}{Region to plot on x-axis.}

\item{region2}{Region to plot on y-axis.}

\item{ratio}{(optional) Name of the ratio being plotted (for labeling).}
}
\description{
Scatterplot of donor CTP ratio between two regions.
}
