% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glmm.R
\name{plot_donor_residual_correlation_heatmap}
\alias{plot_donor_residual_correlation_heatmap}
\title{Generate a heatmap of Spearman correlations between donor-level residuals across
 multiple brain regions for a specific cell type.}
\usage{
plot_donor_residual_correlation_heatmap(
  model,
  regions,
  cell_type,
  brain_region_col = "brain_region_abbreviation_simple",
  donor_col = "donor_external_id"
)
}
\arguments{
\item{model}{Fitted beta-binomial GLMM object for a specific cell type.}

\item{regions}{Brain regions to include, ordered how they should appear in the heatmap.}

\item{cell_type}{Name of the cell type corresponding to the fitted model (for labeling)}

\item{brain_region_col}{Name of the column in the model data that corresponds to brain region.}

\item{donor_col}{Name of the column in the model data that corresponds to donor.}
}
\value{
ggplot object representing the heatmap of Spearman correlations between
donor-level residuals
}
\description{
Generate a heatmap of Spearman correlations between donor-level residuals across
 multiple brain regions for a specific cell type.
}
