% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glmm.R
\name{plot_donor_residual_correlation_scatter}
\alias{plot_donor_residual_correlation_scatter}
\title{Generate a scatter plot comparing donor-level residuals between two brain regions
for a specific cell type, with points colored by age and a linear regression line added.}
\usage{
plot_donor_residual_correlation_scatter(
  model,
  region1,
  region2,
  cell_type,
  brain_region_col = "brain_region_abbreviation_simple",
  donor_col = "donor_external_id",
  age_col = "age_decades"
)
}
\arguments{
\item{model}{Fitted beta-binomial GLMM object for a specific cell type.}

\item{region1}{Name of the first brain region to compare (for x-axis).}

\item{region2}{Name of the second brain region to compare (for y-axis).}

\item{cell_type}{Name of the cell type corresponding to the fitted model (for labeling)}

\item{brain_region_col}{Name of the column in the model data that corresponds to brain region.}

\item{donor_col}{Name of the column in the model data that corresponds to donor.}

\item{age_col}{Name of the column in the model data that corresponds to donor age (for coloring points).}
}
\value{
ggplot object representing the scatter plot comparing donor-level residuals
between the two specified brain regions for the given cell type,
with points colored by age, a linear regression line added, and identity line.
}
\description{
Generate a scatter plot comparing donor-level residuals between two brain regions
for a specific cell type, with points colored by age and a linear regression line added.
}
