% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{plot_variance_partition_results}
\alias{plot_variance_partition_results}
\title{Plots variance explained for each variable across cell types for all the results.}
\usage{
plot_variance_partition_results(combined_results)
}
\arguments{
\item{combined_results}{Dataframe combining variance explained for each variable and cell type across all analyses.}
}
\value{
ggplot of violin and boxplots of variance explained for each variable across cell types.
}
\description{
Plots variance explained for each variable across cell types for all the results.
}
