% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_from_beta_binomial_glmm.R
\name{predict_many_regions}
\alias{predict_many_regions}
\title{Generate predictions from a fitted model for multiple brain regions,
varying age and holding other variables constant.}
\usage{
predict_many_regions(
  model,
  region_names,
  region_col = "brain_region_abbreviation_simple"
)
}
\arguments{
\item{model}{A fitted model object for COUNTS data (e.g., from glmmTMB).}

\item{region_names}{A vector of brain region names to generate predictions for.}

\item{region_col}{Name of the column in the model data that corresponds to brain region.}
}
\value{
A data frame containing the prediction grid with columns for age,
brain region, predicted proportion, and confidence intervals for all specified brain regions.
}
\description{
Generate predictions from a fitted model for multiple brain regions,
varying age and holding other variables constant.
}
