% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_from_beta_binomial_glmm.R
\name{predict_one_region}
\alias{predict_one_region}
\title{Generate predictions from a fitted model for a specific brain region,
varying age and holding other variables constant.}
\usage{
predict_one_region(
  model,
  region_name,
  region_col = "brain_region_abbreviation_simple"
)
}
\arguments{
\item{model}{A fitted model object for COUNTS data (e.g., from glmmTMB).}

\item{region_name}{Name of the brain region to generate predictions for.}

\item{region_col}{Name of the column in the model data that corresponds to brain region}
}
\value{
A data frame containing the prediction grid with columns for age,
brain region, predicted proportion, and standard errors.
}
\description{
Generate predictions from a fitted model for a specific brain region,
varying age and holding other variables constant.
}
