% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glm.R
\name{prepare_data_for_glm}
\alias{prepare_data_for_glm}
\title{Combine cell type proportions and metadata dataframe.}
\usage{
prepare_data_for_glm(sample_ctp, sample_metadata, cell_type_col)
}
\arguments{
\item{sample_ctp}{Dataframe containing cell type proportions for each sample.}

\item{sample_metadata}{Dataframe containing metadata for each sample.}

\item{cell_type_col}{Column name in `sample_ctp` that contains cell type labels.}
}
\value{
Dataframe with combined cell type proportions and metadata, including a column for the number of nuclei of other cell types (total_nuclei - n_nuclei).
}
\description{
Combine cell type proportions and metadata dataframe.
}
