% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{prepare_vp_input}
\alias{prepare_vp_input}
\title{Combines logit-transformed CTP and formatted metadata into a list for variance partition input.}
\usage{
prepare_vp_input(sample_ctp, sample_metadata, cell_type_col, pseudocount = 0.5)
}
\arguments{
\item{sample_ctp}{Dataframe containing cell type proportions for each sample.}

\item{sample_metadata}{Dataframe containing metadata for each sample.}

\item{cell_type_col}{Column name in `sample_ctp` that contains cell type labels.}

\item{pseudocount}{Value to add to zero counts to avoid issues with logit.}
}
\value{
List containing logit-transformed CTP matrix and formatted metadata dataframe, ready for variance partition analysis.
}
\description{
Combines logit-transformed CTP and formatted metadata into a list for variance partition input.
}
