% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{prepare_vp_input_one_cell_type}
\alias{prepare_vp_input_one_cell_type}
\title{Prepares logit-transformed CTP and cell type-specific QC covariates for a single cell type, formatted for variance partition input.}
\usage{
prepare_vp_input_one_cell_type(
  sample_ctp,
  sample_metadata,
  cell_type,
  cell_type_col,
  qc_covariate_cols,
  pseudocount = 0.5
)
}
\arguments{
\item{sample_ctp}{Dataframe containing cell type proportions and QC covariates for each sample.}

\item{sample_metadata}{Dataframe containing metadata for each sample.}

\item{cell_type}{Cell type for which to prepare the variance partition input.}

\item{cell_type_col}{Column name in `sample_ctp` that contains cell type labels.}

\item{qc_covariate_cols}{Character vector of column names in `sample_ctp` that contain the QC covariates to extract and scale for the specified cell type.}

\item{pseudocount}{Value to add to zero counts to avoid issues with logit.}
}
\value{
List containing the logit-transformed CTP matrix for the specified cell type
and the formatted metadata dataframe with cell type-specific QC covariates, ready for variance partition analysis.
}
\description{
Prepares logit-transformed CTP and cell type-specific QC covariates for a single cell type, formatted for variance partition input.
}
