% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ctp.R
\name{regroup_ctp}
\alias{regroup_ctp}
\title{Recompute CTP from an existing CTP dataframe - either with a subset of the
existing grouping variables, or a new cell type label.}
\usage{
regroup_ctp(ctp_df, group_cols, cell_type_col, cell_type_label_map)
}
\arguments{
\item{ctp_df}{Dataframe of cell type proportions.}

\item{group_cols}{Character vector of columns to group by (e.g., donor ID)}

\item{cell_type_col}{Column name representing cell type annotations.}

\item{cell_type_label_map}{Dataframe with 2 columns: the first column should
match the existing cell type labels in the `cell_type_col` of the `ctp_df`,
and the second column should have the new cell type labels to group by.}
}
\value{
Dataframe with recomputed cell type proportions based on the new grouping.
}
\description{
This is useful for quickly recomputing CTP from an existing dataframe
rather than having to load the entire cell metadata file in again. The
tradeoff is that you cannot compute any single-cell metrics.
}
