% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{run_variance_partition}
\alias{run_variance_partition}
\title{Runs variance partition analysis using the `variancePartition` package.}
\usage{
run_variance_partition(vp_input, formula, cell_types = NULL)
}
\arguments{
\item{vp_input}{List containing logit-transformed CTP matrix and formatted metadata dataframe.}

\item{formula}{Formula specifying the fixed and random effects for the variance partition model.}

\item{cell_types}{Optional character vector of cell types to include in the analysis.}
}
\value{
varPartResults object containing the results of the variance partition analysis,
 with variance explained for each variable and cell type.
}
\description{
Runs variance partition analysis using the `variancePartition` package.
}
