% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition.R
\name{run_variance_partition_per_cell_type}
\alias{run_variance_partition_per_cell_type}
\title{Runs variance partition analysis for each specified cell type, incorporating cell type-specific QC covariates, and combines results into a single dataframe.}
\usage{
run_variance_partition_per_cell_type(
  sample_ctp,
  sample_metadata,
  formula,
  cell_types,
  cell_type_col,
  qc_covariate_cols,
  pseudocount = 0.5
)
}
\arguments{
\item{sample_ctp}{Dataframe containing cell type proportions and QC covariates for each sample.}

\item{sample_metadata}{Dataframe containing metadata for each sample.}

\item{formula}{Formula specifying the fixed and random effects for the variance partition model.}

\item{cell_types}{Cell types for which to fit the variance partition model.}

\item{cell_type_col}{Column name in `sample_ctp` that contains cell type labels.}

\item{qc_covariate_cols}{Character vector of column names in `sample_ctp` that contain the QC covariates to extract and scale for the specified cell type.}

\item{pseudocount}{Value to add to zero counts to avoid issues with logit.}
}
\value{
Dataframe combining variance explained for each variable and cell type across all specified cell types.
}
\description{
Runs variance partition analysis for each specified cell type, incorporating cell type-specific QC covariates, and combines results into a single dataframe.
}
