% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ctp.R
\name{save_ctp}
\alias{save_ctp}
\title{Save CTP results.}
\usage{
save_ctp(ctp_df, out_dir, output_prefix)
}
\arguments{
\item{ctp_df}{Dataframe containing cell type proportions and optional metrics.}

\item{out_dir}{Output directory to save the results.}

\item{output_prefix}{Prefix for the output file name.}
}
\value{
None
}
\description{
Save CTP results.
}
