% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ctp.R
\name{compute_ctp_and_metrics}
\alias{compute_ctp_and_metrics}
\title{Main workflow to compute cell type proportions and optional metrics.}
\usage{
compute_ctp_and_metrics(
  df,
  group_cols,
  cell_type_col,
  metric_cols = NULL,
  filters = NULL,
  group_filters = NULL,
  out_file = NULL
)
}
\arguments{
\item{df}{Dataframe containing cell metadata.}

\item{group_cols}{Character vector of columns to group by (e.g., donor ID)}

\item{cell_type_col}{Column name representing cell type annotations.}

\item{metric_cols}{Optional; Character vector of metric columns to compute means for.}

\item{filters}{Optional; Character vector of filtering expressions.}

\item{group_filters}{Optional;  Filtering expressions to apply within groups.}

\item{out_file}{Optional; Output file to save the results}
}
\value{
Dataframe with cell type proportions and optional metrics.
}
\description{
Main workflow to compute cell type proportions and optional metrics.
}
