% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glmm.R
\name{extract_donor_residuals}
\alias{extract_donor_residuals}
\title{Extract donor-level residuals from a fitted beta-binomial GLMM and reshape into a wide format for correlation analysis.}
\usage{
extract_donor_residuals(
  model,
  brain_region_col = "brain_region_abbreviation_simple",
  donor_col = "donor_external_id"
)
}
\arguments{
\item{model}{Fitted beta-binomial GLMM object for a specific cell type.}

\item{brain_region_col}{Name of the column in the model data that corresponds to brain region.}

\item{donor_col}{Name of the column in the model data that corresponds to donor.}
}
\value{
Dataframe containing donor-level residuals for each brain region in wide format,
with one row per donor and columns for each brain region's residuals.
}
\description{
Extract donor-level residuals from a fitted beta-binomial GLMM and reshape into a wide format for correlation analysis.
}
