% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ctp.R
\name{filter_df}
\alias{filter_df}
\title{Filters dataframe.}
\usage{
filter_df(df, filters = NULL)
}
\arguments{
\item{df}{Dataframe to filter.}

\item{filters}{Character vector of filtering expressions.}
}
\value{
Filtered dataframe.
}
\description{
Filters dataframe.
}
