% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glm.R
\name{fit_glmm_and_extract_fixed_effects}
\alias{fit_glmm_and_extract_fixed_effects}
\title{Main workflow to fit beta-binomial GLMMs for multiple cell types and extract fixed effects into a combined dataframe.}
\usage{
fit_glmm_and_extract_fixed_effects(
  sample_ctp,
  sample_metadata,
  cell_types,
  cell_type_col,
  fixed_effects,
  random_effects,
  out_file = NULL
)
}
\arguments{
\item{sample_ctp}{Dataframe containing cell type proportions for each sample.}

\item{sample_metadata}{Dataframe containing metadata for each sample.}

\item{cell_types}{Cell types for which to fit models and extract fixed effects.}

\item{cell_type_col}{Column name in `sample_df` that contains cell type labels, and to use for the cell type column in the output fixed effects dataframe.}

\item{fixed_effects}{Character vector of column names in `sample_df` to include as fixed effects in the models.}

\item{random_effects}{Character vector of column names in `sample_df` to include as random effects in the models.}

\item{out_file}{Optional file path to save the combined fixed effects dataframe as a tab-delimited text file.}
}
\value{
List containing the named list of fitted beta-binomial GLMM objects for each cell type and the combined fixed effects dataframe for all models.
}
\description{
Main workflow to fit beta-binomial GLMMs for multiple cell types and extract fixed effects into a combined dataframe.
}
