% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_beta_binomial_glmm.R
\name{fixed_effect_volcano_plot}
\alias{fixed_effect_volcano_plot}
\title{Generate a volcano plot for a specific fixed effect from the combined fixed effects dataframe.}
\usage{
fixed_effect_volcano_plot(fixed_effects_df, fixed_effect)
}
\arguments{
\item{fixed_effects_df}{Dataframe containing the fixed effects estimates, confidence intervals, and cell type information for all specified models.}

\item{fixed_effect}{Name of the fixed effect for which to generate the volcano plot}
}
\value{
ggplot object representing the volcano plot for the specified fixed effect,
showing effect size (log-odds) on the x-axis and -log10(p-value) on the y-axis,
with points colored by cell type and error bars representing confidence intervals.
}
\description{
Generate a volcano plot for a specific fixed effect from the combined fixed effects dataframe.
}
