% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_ctp_outliers.R
\name{identify_low_nuclei_count_outliers}
\alias{identify_low_nuclei_count_outliers}
\title{Identify samples to drop based on low nuclei counts.}
\usage{
identify_low_nuclei_count_outliers(ctp_df)
}
\arguments{
\item{ctp_df}{Dataframe of cell type counts and proportions (one row per cell type).}
}
\value{
Dataframe with sample_id, total_nuclei, z_log10_total_nuclei, and nuclei_count_outlier (TRUE/FALSE).
}
\description{
Identify samples to drop based on low nuclei counts.
}
