% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctp_correlations.R
\name{make_donor_ctp_wide}
\alias{make_donor_ctp_wide}
\title{Make a wide-format dataframe of donor cell type proportions, with columns
for each cell type or brain region.}
\usage{
make_donor_ctp_wide(
  ctp_df,
  filter_value = NULL,
  filter_col = NULL,
  pivot_cols = NULL,
  cell_type_col = "annotation",
  region_col = "brain_region_abbreviation_simple",
  donor_col = "donor_external_id",
  metric_col = "fraction_nuclei",
  drop_outliers = TRUE
)
}
\arguments{
\item{ctp_df}{Dataframe containing cell type proportions, with columns for
cell type, brain region, donor ID, and the CTP metric.}

\item{filter_value}{Optional value to filter the dataframe by
(e.g., a specific cell type or brain region).}

\item{filter_col}{Column of dataframe to apply the filter on (e.g., cell type
column or brain region column).}

\item{cell_type_col}{Column of dataframe containing cell type labels.}

\item{region_col}{Column of dataframe containing brain region labels.}

\item{donor_col}{Column of dataframe containing donor IDs.}

\item{metric_col}{Column of dataframe containing the CTP metric to be plotted (e.g., fraction of nuclei).}

\item{drop_outliers}{Whether to filter out rows marked as outliers
(if an "outlier" column exists in the dataframe). Default is TRUE.}

\item{pivot_col}{Column to pivot on to create wide format (e.g., cell type
column or brain region column).}
}
\value{
A wide-format dataframe with rows for each donor and columns for each
cell type or brain region (depending on the pivot_col), containing the specified CTP metric.
}
\description{
Make a wide-format dataframe of donor cell type proportions, with columns
for each cell type or brain region.
}
